Fix fmt violation
The newest version of fmt enables more safety flags to ensure that no
implicit casts are done by default.  The line changed was trying to
implicitly cast an enum.  This commit makes this explicit with a
static_cast.

Tested: Code compiles against the latest subtree update.

+++ monitor/tach_sensor.cpp
@@ -295,7 +295,7 @@ void TachSensor::startTimer(TimerMode mode)
     {
         log<level::DEBUG>(
             fmt::format("Start timer({}) on tach sensor {}. [delay = {}s]",
-                        mode, _name,
+                        static_cast<int>(mode), _name,
                         duration_cast<seconds>(getDelay(mode)).count())
                 .c_str());
         _timer.restartOnce(getDelay(mode));
